docker file simplified
Installing python from source is a real effort. The docker container
takes ages to build and then installing pip from get-pip.py seems to
require and ssl package which is quite mysterious to me. while the
installed python version itself is supposed to have pip, thats not the
case. I tried to install python-pip from ubuntu to the result that
installed packages are not imported in script. This left me with the
trivial option of installing both python and pip from ubuntu repository
and hoping that it all works.